requirements.txt
Django==6.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
django-cors-headers==4.9.0
django-silk==5.4.3
django-filter>=2.4.0
django-environ==0.12.0
psycopg2-binary==2.9.11
pillow==12.0.0
djangorestframework-simplejwt==5.5.1